use int64 for ts

+++ benchmark/nexmark/pkg/nexmark/generator/bid_generator.go
@@ -28,7 +28,7 @@ type ChannelUrl struct {
 	url     string
 }
 
-func NextBid(ctx context.Context, eventId uint64, random *rand.Rand, timestamp uint64, config *GeneratorConfig, bidUrlCache map[uint32]*ChannelUrl) (*types.Bid, error) {
+func NextBid(ctx context.Context, eventId uint64, random *rand.Rand, timestamp int64, config *GeneratorConfig, bidUrlCache map[uint32]*ChannelUrl) (*types.Bid, error) {
 	auction := uint64(0)
 	if random.Intn(int(config.Configuration.HotAuctionRatio)) > 0 {
 		auction = LastBase0AuctionId(config, eventId) / uint64(HOT_AUCTION_RATIO) * uint64(HOT_AUCTION_RATIO)
@@ -80,7 +80,7 @@ func NextBid(ctx context.Context, eventId uint64, random *rand.Rand, timestamp u
 		Price:    price,
 		Channel:  channel,
 		Url:      url,
-		DateTime: int64(timestamp),
+		DateTime: timestamp,
 		Extra:    extra,
 	}, nil
 }
